# openapi.yaml
openapi: 3.1.0
info:
  title: Wave Billing GPT Backend
  version: "0.1.0"
  description: |
    Backend for Wave + Custom GPT integrations. All business-facing endpoints are
    protected by an internal header `x-internal-secret` and communicate with Wave's
    GraphQL API. Use these endpoints as OpenAI Actions.
servers:
  - url: https://wave-billing-gpt-backend.vercel.app
    description: Production server (replace with real Vercel URL)
security:
  - InternalApiKey: []
components:
  securitySchemes:
    InternalApiKey:
      type: apiKey
      in: header
      name: x-internal-secret
  schemas:
    BusinessKey:
      type: string
      enum: [manna, bako, socialion]
      description: Logical business identifier mapped to Wave business IDs.
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: array
          items:
            type: string
      required: [error]
...
    CreateInvoiceItemInput:
      type: object
      properties:
        description:
          type: string
        unitPrice:
          type: number
        quantity:
          type: number
      required: [description, unitPrice, quantity]
    CreateInvoiceCustomerInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
      required: [name]
    CreateInvoiceRequestBody:
      type: object
      properties:
        businessKey:
          $ref: '#/components/schemas/BusinessKey'
        customer:
          $ref: '#/components/schemas/CreateInvoiceCustomerInput'
        items:
          type: array
          items:
            $ref: '#/components/schemas/CreateInvoiceItemInput'
          minItems: 1
        invoiceDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        currencyCode:
          type: string
          default: USD
        notes:
          type: string
      required: [businessKey, customer, items]
    CreatedInvoiceCustomer:
      type: object
      properties:
        id:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
      required: [id, name, email]
    CreatedInvoicePayload:
      type: object
      properties:
        id:
          type: string
        invoiceNumber:
          type: string
          nullable: true
        status:
          type: string
        createdAt:
          type: string
        dueDate:
          type: string
          nullable: true
        totalAmount:
          type: number
        totalCurrency:
          type: string
        amountDue:
          type: number
        customer:
          $ref: '#/components/schemas/CreatedInvoiceCustomer'
        publicUrl:
          type: string
          nullable: true
      required: [id, status, createdAt, totalAmount, totalCurrency, amountDue, customer]
    CreatedInvoiceResponseBody:
      type: object
      properties:
        businessKey:
          $ref: '#/components/schemas/BusinessKey'
        businessId:
          type: string
        invoice:
          $ref: '#/components/schemas/CreatedInvoicePayload'
      required: [businessKey, businessId, invoice]
...
paths:
...
  /api/wave/invoices/create:
    post:
      summary: Create a new invoice in Wave
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvoiceRequestBody'
      responses:
        '200':
          description: Invoice created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedInvoiceResponseBody'
        '400':
          description: Validation or Wave input error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected creation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...